refactor(config): decompose config.py into apts/config/ package
Split the monolithic apts/config.py into a structured package with
specialized submodules: base, display, weather, cache, performance,
data, place, api, events, and objects. Preserved the public API
via re-exports in apts/config/__init__.py and maintained logger
identity for test compatibility.
